+++ src/components/AnchorForAttachmentsOnly/index.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import * as anchorForAttachmentsOnlyPropTypes from './anchorForAttachmentsOnlyPropTypes';
+import BaseAnchorForAttachmentsOnly from './BaseAnchorForAttachmentsOnly';
+
+const AnchorForAttachmentsOnly = props => <BaseAnchorForAttachmentsOnly {...props} />;
+
+AnchorForAttachmentsOnly.propTypes = anchorForAttachmentsOnlyPropTypes.propTypes;
+AnchorForAttachmentsOnly.defaultProps = anchorForAttachmentsOnlyPropTypes.defaultProps;
+AnchorForAttachmentsOnly.displayName = 'AnchorForAttachmentsOnly';
+
+export default AnchorForAttachmentsOnly;
